Time: 800 ms (5.77%) | Memory: 54.9 MB (6.73%) - LeetSync

+++ 1800-concatenation-of-consecutive-binary-numbers/concatenation-of-consecutive-binary-numbers.java
@@ -15,6 +15,6 @@ class Solution {
             pow2=(pow2%MOD)*2;
             concatInt%=MOD;
         }
-        return (int)concatInt;
+        return concatInt;
     }
 }
